Remove special case output to STDOUT
The spec formatter checks that the output is a tty to decide whether to
use colourized formatting but then bypasses this check if STDOUT is used.
This means spec > output.log has undesirable escape sequences.  The only
purpose for this special casing seems to be preventing the closure of
STDOUT.  Localize this special case to #close so #output_to_tty? works
as expected.