Removed example_group_description
It had been added as part of [#660], but the problem in that ticket was
instances of example groups (the context in which examples are run) and
the instance variables in those instances. Forcing the class out of
scope also was serving little benefit but adding unnecessary cost
(looping through the hierarchy for the nested formatter even when using
other formatters that don't need it).